requirements.txt
fastapi==0.128.0
langchain==1.2.0
langchain-community==0.4.1
langchain-openai==1.1.6
langchain-qdrant==1.1.0
pypdf==6.5.0
qdrant-client==1.16.2
openai==2.14.0
python-dotenv==1.2.1
tiktoken==0.12.0